+++ src/main/java/kr/co/csalgo/web/subscription/dto/EmailVerificationVerifyDto.java
@@ -1,16 +1,18 @@
-package kr.co.csalgo.web.subscribe.dto;
+package kr.co.csalgo.web.subscription.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @AllArgsConstructor
-public class SubscriptionDto {
+public class EmailVerificationVerifyDto {
 	@NoArgsConstructor
 	@AllArgsConstructor
 	@Getter
 	public static class Request {
 		private String email;
+		private String code;
+		private VerificationCodeType type;
 	}
 
 	@NoArgsConstructor
