+++ src/main/java/kr/co/csalgo/web/subscription/dto/SubscriptionDto.java
@@ -1,16 +1,16 @@
-package kr.co.csalgo.web.subscribe.dto;
+package kr.co.csalgo.web.subscription.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-public class EmailVerificationDto {
+@AllArgsConstructor
+public class SubscriptionDto {
 	@NoArgsConstructor
 	@AllArgsConstructor
 	@Getter
 	public static class Request {
 		private String email;
-		private VerificationCodeType type;
 	}
 
 	@NoArgsConstructor
