+++ src/main/java/kr/co/csalgo/web/subscription/dto/VerificationCodeType.java
@@ -1,4 +1,4 @@
-package kr.co.csalgo.web.subscribe.dto;
+package kr.co.csalgo.web.subscription.dto;
 
 public enum VerificationCodeType {
 	SUBSCRIPTION
