+++ src/main/resources/templates/fragments/common/navbar.html
@@ -16,5 +16,5 @@
             </div>
         </div>
     </nav>
-    <div th:replace="fragments/modal/subscribe-modal :: subscribeModal"></div>
+    <div th:replace="fragments/modal/subscription-modal :: subscriptionModal"></div>
 </div>
