+++ src/main/resources/templates/fragments/modal/subscription-modal.html
@@ -1,15 +1,15 @@
-<div th:fragment="subscribeModal">
-    <div class="modal fade" id="subscribeModal" tabindex="-1" aria-labelledby="subscribeModalLabel" aria-hidden="true">
+<div th:fragment="subscriptionModal">
+    <div class="modal fade" id="subscriptionModal" tabindex="-1" aria-labelledby="subscriptionModalLabel" aria-hidden="true">
         <div class="modal-dialog modal-dialog-centered">
             <div class="modal-content">
 
                 <div class="modal-header">
-                    <h5 class="modal-title" id="subscribeModalLabel">이메일 인증</h5>
+                    <h5 class="modal-title" id="subscriptionModalLabel">이메일 인증</h5>
                     <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="닫기"></button>
                 </div>
 
                 <div class="modal-body">
-                    <form id="subscribeForm">
+                    <form id="subscriptionForm">
                         <div class="mb-3">
                             <label for="email" class="form-label">이메일 주소</label>
                             <input type="email" class="form-control" id="email" name="email" required>
@@ -40,7 +40,7 @@
 
                 <div class="modal-footer">
                     <button type="button" class="btn btn-secondary" data-bs-dismiss="modal">닫기</button>
-                    <button type="submit" class="btn btn-primary" form="subscribeForm" id="submitBtn" disabled>
+                    <button type="submit" class="btn btn-primary" form="subscriptionForm" id="submitBtn" disabled>
                         확인
                     </button>
                 </div>
