+++ src/main/resources/templates/index.html
@@ -10,14 +10,14 @@
           href="https://cdn.jsdelivr.net/gh/orioncactus/pretendard@v1.3.9/dist/web/variable/pretendardvariable.min.css"/>
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css" rel="stylesheet">
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/js/bootstrap.bundle.min.js"></script>
-    <script src="/js/subscribe-modal.js"></script>
+    <script type="module" src="/js/subscription-modal.js"></script>
 </head>
 <body>
 <div th:replace="fragments/common/navbar :: navbar"></div>
 <section id="intro" class="bg-white text-center">
     <div class="section-content">
         <h2>서비스 소개 문구</h2>
-        <button type="button" class="btn btn-primary mt-3" data-bs-toggle="modal" data-bs-target="#subscribeModal">
+        <button type="button" class="btn btn-primary mt-3" data-bs-toggle="modal" data-bs-target="#subscriptionModal">
             지금 시작하기
         </button>
     </div>
@@ -65,6 +65,6 @@
     </div>
 </section>
 <div th:replace="fragments/common/footer :: footer"></div>
-<div th:replace="fragments/modal/subscribe-modal :: subscribeModal"></div>
+<div th:replace="fragments/modal/subscription-modal :: subscriptionModal"></div>
 </body>
 </html>
